+++ src/pages/mypage/MyPage.jsx
@@ -1,5 +1,20 @@
+import Button from '@components/ui/Button';
+import useMemberStore from '@zustand/memberStore.mjs';
+import { useNavigate } from 'react-router-dom';
+
 function MyPage() {
-  return <div>MyPage</div>;
+  const navigate = useNavigate();
+  const setUserData = useMemberStore((state) => state.setUser);
+
+  const handleLogout = () => {
+    setUserData(null);
+    navigate('/');
+  };
+  return (
+    <div>
+      <Button onClick={handleLogout}>로그아웃</Button>
+    </div>
+  );
 }
 
 export default MyPage;
